refactor the project, adding an infra folder for the project's technical implementations that are outside the scope of the project's business rules
