+++ src/modules/accounts/infra/typeorm/repositories/UsersRepository.ts
@@ -1,8 +1,13 @@
 import { getRepository, Repository } from 'typeorm';
 
-import { ICreateUserDTO } from '../../dtos/ICreateUserDTO';
-import { User } from '../../entities/User';
-import { IUsersRepository } from '../IUsersRepository';
+import { ICreateUserDTO } from '@modules/accounts/dtos/ICreateUserDTO';
+import { IUsersRepository } from '@modules/accounts/repositories/IUsersRepository';
+
+import { User } from '../entities/User';
+
+// import { ICreateUserDTO } from '../../dtos/ICreateUserDTO';
+// import { User } from '../../infra/typeorm/entities/User';
+// import { IUsersRepository } from '../IUsersRepository';
 
 class UsersRepository implements IUsersRepository {
   private repository: Repository<User>;
