+++ src/modules/accounts/useCases/authenticateUser/AuthenticateUserUseCase.ts
@@ -3,9 +3,9 @@ import { sign } from 'jsonwebtoken';
 import { inject, injectable } from 'tsyringe';
 
 // import { AppError } from '../../../../errors/AppError';
-import { AppError } from '@errors/AppError';
-// import { IUsersRepository } from '../../repositories/IUsersRepository';
 import { IUsersRepository } from '@modules/accounts/repositories/IUsersRepository';
+import { AppError } from '@shared/errors/AppError';
+// import { IUsersRepository } from '../../repositories/IUsersRepository';
 
 interface IRequest {
   email: string;
