+++ src/modules/accounts/useCases/createUser/CreateUserUseCase.ts
@@ -2,9 +2,9 @@ import { hash } from 'bcryptjs';
 import { inject, injectable } from 'tsyringe';
 
 // import { AppError } from '../../../../errors/AppError';
-import { AppError } from '@errors/AppError';
 import { ICreateUserDTO } from '@modules/accounts/dtos/ICreateUserDTO';
 import { IUsersRepository } from '@modules/accounts/repositories/IUsersRepository';
+import { AppError } from '@shared/errors/AppError';
 
 // import {
 //   ICreateUserDTO,
