+++ src/modules/cars/infra/typeorm/repositories/CategoriesRepository.ts
@@ -1,10 +1,17 @@
 import { getRepository, Repository } from 'typeorm';
 
-import { Category } from '../../entities/Category';
 import {
   ICategoriesRepository,
   ICreateCategoryDTO,
-} from '../ICategoriesRepository';
+} from '@modules/cars/repositories/ICategoriesRepository';
+
+import { Category } from '../entities/Category';
+
+// import { Category } from '../../infra/typeorm/entities/Category';
+// import {
+//   ICategoriesRepository,
+//   ICreateCategoryDTO,
+// } from '../ICategoriesRepository';
 
 class CategoriesRepository implements ICategoriesRepository {
   private repository: Repository<Category>;
