+++ src/modules/cars/infra/typeorm/repositories/SpecificationsRepository.ts
@@ -1,10 +1,17 @@
 import { getRepository, Repository } from 'typeorm';
 
-import { Specification } from '../../entities/Specification';
 import {
   ICreateSpecificationDTO,
   ISpecificationsRepository,
-} from '../ISpecificationsRepository';
+} from '@modules/cars/repositories/ISpecificationsRepository';
+
+import { Specification } from '../entities/Specification';
+
+// import { Specification } from '../../infra/typeorm/entities/Specification';
+// import {
+//   ICreateSpecificationDTO,
+//   ISpecificationsRepository,
+// } from '../ISpecificationsRepository';
 
 class SpecificationsRepository implements ISpecificationsRepository {
   private repository: Repository<Specification>;
