+++ src/modules/cars/useCases/createCategory/CreateCategoryUseCase.spec.ts
@@ -1,7 +1,7 @@
 // import { AppError } from '../../../../errors/AppError';
-import { AppError } from '@errors/AppError';
-// import { CategoriesRepositoryInMemory } from '../../repositories/inMemory/CategoriesRepositoryInMemory';
 import { CategoriesRepositoryInMemory } from '@modules/cars/repositories/inMemory/CategoriesRepositoryInMemory';
+import { AppError } from '@shared/errors/AppError';
+// import { CategoriesRepositoryInMemory } from '../../repositories/inMemory/CategoriesRepositoryInMemory';
 
 import { CreateCategoryUseCase } from './CreateCategoryUseCase';
