+++ src/modules/cars/useCases/createCategory/CreateCategoryUseCase.ts
@@ -1,9 +1,9 @@
 import { inject, injectable } from 'tsyringe';
 
 // import { AppError } from '../../../../errors/AppError';
-import { AppError } from '@errors/AppError';
-// import { ICategoriesRepository } from '../../repositories/ICategoriesRepository';
 import { ICategoriesRepository } from '@modules/cars/repositories/ICategoriesRepository';
+import { AppError } from '@shared/errors/AppError';
+// import { ICategoriesRepository } from '../../repositories/ICategoriesRepository';
 
 interface IRequest {
   name: string;
