+++ src/modules/cars/useCases/createSpecification/CreateSpecificationUseCase.ts
@@ -1,9 +1,9 @@
 import { inject, injectable } from 'tsyringe';
 
 // import { AppError } from '../../../../errors/AppError';
-import { AppError } from '@errors/AppError';
-// import { ISpecificationsRepository } from '../../repositories/ISpecificationsRepository';
 import { ISpecificationsRepository } from '@modules/cars/repositories/ISpecificationsRepository';
+import { AppError } from '@shared/errors/AppError';
+// import { ISpecificationsRepository } from '../../repositories/ISpecificationsRepository';
 
 interface IRequest {
   name: string;
