+++ src/modules/cars/useCases/importCategory/ImportCategoryUseCase.ts
@@ -2,8 +2,10 @@ import { parse as csvParse } from 'csv-parse';
 import fs from 'node:fs';
 import { inject, injectable } from 'tsyringe';
 
+import { CategoriesRepository } from '@modules/cars/infra/typeorm/repositories/CategoriesRepository';
+
 // import { CategoriesRepository } from '../../repositories/implementations/CategoriesRepository';
-import { CategoriesRepository } from '@modules/cars/repositories/implementations/CategoriesRepository';
+// import { CategoriesRepository } from '@modules/cars/repositories/implementations/CategoriesRepository';
 
 interface IImportCategory {
   name: string;
