+++ src/modules/cars/useCases/listCategories/ListCategoriesUseCase.ts
@@ -1,7 +1,7 @@
 import { inject, injectable } from 'tsyringe';
 
 // import { Category } from '../../entities/Category';
-import { Category } from '@modules/cars/entities/Category';
+import { Category } from '@modules/cars/infra/typeorm/entities/Category';
 // import { ICategoriesRepository } from '../../repositories/ICategoriesRepository';
 import { ICategoriesRepository } from '@modules/cars/repositories/ICategoriesRepository';
