+++ src/modules/cars/useCases/listSpecifications/ListSpecificationsUseCase.ts
@@ -1,7 +1,7 @@
 import { inject, injectable } from 'tsyringe';
 
 // import { Specification } from '../../entities/Specification';
-import { Specification } from '@modules/cars/entities/Specification';
+import { Specification } from '@modules/cars/infra/typeorm/entities/Specification';
 // import { ISpecificationsRepository } from '../../repositories/ISpecificationsRepository';
 import { ISpecificationsRepository } from '@modules/cars/repositories/ISpecificationsRepository';
