--- src/routes/specifications.routes.ts
@@ -1,18 +1,0 @@
-import { Router } from 'express';
-
-import { ensureAuthenticated } from '../middlewares/ensureAuthenticated';
-import { CreateSpecificationController } from '../modules/cars/useCases/createSpecification/CreateSpecificationController';
-import { ListSpecificationsController } from '../modules/cars/useCases/listSpecifications/ListSpecificationsController';
-
-const specificationsRoutes = Router();
-
-const createSpecificationController = new CreateSpecificationController();
-const listSpecificationsController = new ListSpecificationsController();
-
-specificationsRoutes.use(ensureAuthenticated);
-
-specificationsRoutes.post('/', createSpecificationController.handle);
-
-specificationsRoutes.get('/', listSpecificationsController.handle);
-
-export { specificationsRoutes };
