--- src/routes/users.routes.ts
@@ -1,24 +1,0 @@
-import { Router } from 'express';
-import multer from 'multer';
-
-import uploadConfig from '../config/upload';
-import { ensureAuthenticated } from '../middlewares/ensureAuthenticated';
-import { CreateUserController } from '../modules/accounts/useCases/createUser/CreateUserController';
-import { UpdateUserAvatarController } from '../modules/accounts/useCases/updateUserAvatar/UpdateUserAvatarController';
-
-const usersRoutes = Router();
-
-const uploadAvatar = multer(uploadConfig.upload('./tmp/avatar'));
-
-const createUserController = new CreateUserController();
-const updateUserAvatarController = new UpdateUserAvatarController();
-
-usersRoutes.post('/', createUserController.handle);
-usersRoutes.patch(
-  '/avatar',
-  ensureAuthenticated,
-  uploadAvatar.single('avatar'),
-  updateUserAvatarController.handle,
-);
-
-export { usersRoutes };
