+++ src/shared/infra/http/middlewares/ensureAuthenticated.ts
@@ -1,8 +1,11 @@
 import { Request, Response, NextFunction, request } from 'express';
 import { verify } from 'jsonwebtoken';
 
-import { AppError } from '../../../../errors/AppError';
-import { UsersRepository } from '../modules/accounts/repositories/implementations/UsersRepository';
+import { UsersRepository } from '@modules/accounts/infra/typeorm/repositories/UsersRepository';
+import { AppError } from '@shared/errors/AppError';
+
+// import { AppError } from '../../../../errors/AppError';
+// import { UsersRepository } from '../modules/accounts/repositories/implementations/UsersRepository';
 
 interface IPayload {
   sub: string;
