+++ src/shared/infra/http/routes/authenticate.routes.ts
@@ -1,6 +1,8 @@
 import { Router } from 'express';
 
-import { AuthenticateUserController } from '../modules/accounts/useCases/authenticateUser/AuthenticateUserController';
+import { AuthenticateUserController } from '@modules/accounts/useCases/authenticateUser/AuthenticateUserController';
+
+// import { AuthenticateUserController } from '../modules/accounts/useCases/authenticateUser/AuthenticateUserController';
 
 const authenticateRoutes = Router();
