+++ src/shared/infra/http/routes/specifications.routes.ts
@@ -1,0 +1,22 @@
+import { Router } from 'express';
+
+import { CreateSpecificationController } from '@modules/cars/useCases/createSpecification/CreateSpecificationController';
+import { ListSpecificationsController } from '@modules/cars/useCases/listSpecifications/ListSpecificationsController';
+import { ensureAuthenticated } from '@shared/infra/http/middlewares/ensureAuthenticated';
+
+// import { ensureAuthenticated } from '../shared/infra/http/middlewares/ensureAuthenticated';
+// import { CreateSpecificationController } from '../modules/cars/useCases/createSpecification/CreateSpecificationController';
+// import { ListSpecificationsController } from '../modules/cars/useCases/listSpecifications/ListSpecificationsController';
+
+const specificationsRoutes = Router();
+
+const createSpecificationController = new CreateSpecificationController();
+const listSpecificationsController = new ListSpecificationsController();
+
+specificationsRoutes.use(ensureAuthenticated);
+
+specificationsRoutes.post('/', createSpecificationController.handle);
+
+specificationsRoutes.get('/', listSpecificationsController.handle);
+
+export { specificationsRoutes };
